+++ src/Env/Util/Loader/EnvLoader.php
@@ -4,36 +4,55 @@ declare(strict_types=1);
 
 namespace LDL\Env\Util\Loader;
 
+use LDL\Env\Util\EnvUtil;
 use LDL\Env\Util\File\Exception\ReadEnvFileException;
 use LDL\Env\Util\File\Parser\EnvFileParser;
 use LDL\Env\Util\Line\Collection\EnvLineCollectionInterface;
 use LDL\Env\Util\Line\Type\Variable\EnvLineVarInterface;
-use LDL\Framework\Helper\IterableHelper;
+use LDL\Type\Collection\Interfaces\Type\StringCollectionInterface;
+use LDL\Type\Collection\Types\String\StringCollection;
 
 final class EnvLoader
 {
+    /**
+     * @var string[]
+     */
+    private static $loaded = [];
+
     public static function load(EnvLineCollectionInterface $lines): void
     {
-        $lines = IterableHelper::filter($lines, static function ($l) {
-            return $l instanceof EnvLineVarInterface;
-        });
-
-        IterableHelper::map($lines, static function (EnvLineVarInterface $l) {
-            putenv((string) $l);
-            $_ENV[$l->getVar(true)] = $l->getValue();
-        });
+        foreach ($lines as $line) {
+            if ($line instanceof EnvLineVarInterface) {
+                EnvUtil::set($line->getVar(true), $line->getValue());
+            }
+        }
     }
 
     /**
-     * This method must be used with an .env file which has already been compiled.
+     * Loads a .env file which is already compiled (in the form of VAR=VALUE)
+     * NOTE: This method must ONLY be used with a regular .env file which has already been compiled.
      *
      * @throws ReadEnvFileException
      */
     public static function loadFile(string $file): void
     {
-        self::load(
-            (new EnvFileParser(null, false))
-                ->parse([$file], false)
-        );
+        try {
+            self::load(
+                (new EnvFileParser(null, false))
+                    ->parse([$file], false)
+            );
+
+            self::$loaded[$file] = true;
+        } catch (\Throwable $e) {
+            throw $e;
+        }
+    }
+
+    /**
+     * Returns a string collection of file paths loaded by self::loadFile.
+     */
+    public static function getLoadedFiles(): StringCollectionInterface
+    {
+        return new StringCollection(self::$loaded);
     }
 }